Explicitly prevent fits with fewer data points than parameters
There can be situations where numerical errors cause the matrix to be
not strictly singular.  The results are of course invalid, in that,
while the fit should indeed pass through all data points, the
parameters are one of infinitely-many possible choices, and the
finite-variance statistics computed with Monte Carlo sampling do not
reflect this.

Also replaced calls to dsytr{f,i} with calls to dgetr{f,i} (not much
of a performance penalty, saves symmetrizing step).