Implemented "qrandom" variable.
This facility estimates the quasirandom error in the data, modelled
by sigma=alpha*X^gamma, and modifies the standard error to incorporate
this.  Thus, extrapolations account for the quasirandom error.  The
value of gamma is supplied by the user with "set qrandom <gamma>",
and the value of gamma is determined before any Monte Carlo resample
by analyzing the deviations of the data from a non-resampled fit.