Implemented an "intersect" mode for correlated qrandom noise.
In this method, qrandom "alpha" prefactors are evaluated for all sets,
and a linear combination z_ij = y_i + beta*(y_j-y_i) of each pair of
datasets is constructed such that (a) correlated quasirandom noise is
not present in z_ij, and (b) z_ij = y_i = y_j at the intersection.

All pairs of datasets, including all {y_i} and all {z_ij}, are
intersected, and the intersection yielding the smallest dy0 is
reported as "INTR BEST".

This method requires qrandom to be set and all datasets to share the
same x values.

In a test on synthetic data, the bias in y-y intersections and y-yhat
intersections are similar, and both are about 3 orders of magnitude
greater than the bias in yhat-yhat intersections.