Made faster version of perform_multifit for cases without shared parameters.
The shared-parameter version of perform_multifit scales quadratically
with the number of sets, while the new version is linear.  For a
handful of sets with a handful of parameters this is usually not a
problem, but I'm testing functionality which spawns sets by the
hundreds and the unified approach starts to show its sluggishness.